updated rain; implemented fog while raining;

+++ MAC420_ComputacaoGrafica/EP2/src/engine/color.h
@@ -7,14 +7,14 @@ namespace engine {
 class Color {
   public:
 
-    Color() : r(1.0), g(1.0), b(1.0), a(1.0) {}
-    Color(double _r, double _g, double _b, double _a = 1.0) : r(_r), g(_g), b(_b), a(_a) {}
+    Color() : r(1.0f), g(1.0f), b(1.0f), a(1.0f) {}
+    Color(float _r, float _g, float _b, float _a = 1.0) : r(_r), g(_g), b(_b), a(_a) {}
 
     ~Color() { }
 
     union {
-        struct { double r, g, b, a; };
-        struct { double val[4]; };
+        struct { float r, g, b, a; };
+        struct { float val[4]; };
     };
 
 };
